+++ src/libawkward/Identities.cpp
@@ -72,7 +72,7 @@ namespace awkward {
                                 kernel::Lib ptr_lib)
       : Identities(ref, fieldloc, 0, width, length)
       , ptr_lib_(ptr_lib)
-      , ptr_(kernel::ptr_alloc<T>(ptr_lib, (ssize_t)width*length)) { }
+      , ptr_(kernel::ptr_alloc<T>(ptr_lib, width*length)) { }
 
   template <typename T>
   IdentitiesOf<T>::IdentitiesOf(const Ref ref,
@@ -349,7 +349,7 @@ namespace awkward {
                                                ptr_,
                                                ptr_lib_);
     }
-    std::shared_ptr<T> ptr = kernel::ptr_alloc<T>(ptr_lib, width_ * length_);
+    std::shared_ptr<T> ptr = kernel::ptr_alloc<T>(ptr_lib, width_*length_);
 
     Error err =  kernel::copy_to<T>(ptr_lib,
                                     ptr_lib_,
